Prepare for scheduling, temporary setpoint override.
Add parsing for CiA 301 TIME_OF_DAY structs. These values
are used to store the end time of the temporary setpoint override.
