+++ custom_components/remeha_modbus/api/__init__.py
@@ -3,16 +3,18 @@
 __all__ = []
 
 from .api import (
-    ClimateZone,  # noqa: F401
-    ClimateZoneFunction,  # noqa: F401
-    ClimateZoneHeatingMode,  # noqa: F401
-    ClimateZoneMode,  # noqa: F401
-    ClimateZoneScheduleId,  # noqa: F401
-    ClimateZoneType,  # noqa: F401
     ConnectionType,  # noqa: F401
     DeviceBoardCategory,  # noqa: F401
     DeviceBoardType,  # noqa: F401
     DeviceInstance,  # noqa: F401
     RemehaApi,  # noqa: F401
     SerialConnectionMethod,  # noqa: F401
+)
+from .climate_zone import (
+    ClimateZone,  # noqa: F401
+    ClimateZoneFunction,  # noqa: F401
+    ClimateZoneHeatingMode,  # noqa: F401
+    ClimateZoneMode,  # noqa: F401
+    ClimateZoneScheduleId,  # noqa: F401
+    ClimateZoneType,  # noqa: F401
 )
