+++ custom_components/remeha_modbus/manifest.json
@@ -11,7 +11,7 @@
   "loggers": ["remeha_modbus"],
   "quality_scale": "bronze",
   "requirements": [
-    "pymodbus==3.8.3",
+    "pymodbus==3.9.2",
     "pydantic==2.10.6",
     "python-dateutil==2.9.0.post0"
   ],
