+++ requirements.test.txt
@@ -4,5 +4,5 @@ pytest
 pytest-cov
 pytest-homeassistant-custom-component
 pytest-asyncio==0.26.0
-pymodbus==3.8.3
+pymodbus==3.9.2
 ruff==0.11.5
