+++ tests/fixtures/modbus_store.json
@@ -22,11 +22,15 @@
       "644": "4131",
       "646": "0001",
       "649": "0001",
+      "663": "ff00",
       "664": "00c8",
       "665": "ff00",
       "666": "ff00",
       "686": "ff00",
       "688": "0000",
+      "978": "ff00",
+      "979": "ff00",
+      "980": "ff00",
       "1104": "00e8",
       "1109": "0002",
       "1110": "0001",
@@ -38,11 +42,15 @@
       "1156": "0000",
       "1158": "0001",
       "1161": "0001",
+      "1175": "ff00",
       "1176": "ff00",
       "1177": "157c",
       "1178": "09c4",
       "1198": "012c",
       "1200": "0001",
+      "1490": "ff00",
+      "1491": "ff00",
+      "1492": "ff00",
       "1616": "0080",
       "1621": "0000",
       "1622": "0000",
