+++ tests/fixtures/modbus_store_ch_scheduling.json
@@ -21,13 +21,24 @@
       "643": "5243",
       "644": "4131",
       "646": "0001",
-      "649": "0001",
+      "649": "0000",
       "663": "ff00",
       "664": "00c8",
       "665": "ff00",
       "666": "ff00",
       "686": "ff00",
       "688": "0000",
+      "689": "0510",
+      "690": "c824",
+      "691": "30c8",
+      "692": "2a20",
+      "693": "c854",
+      "694": "40c8",
+      "695": "60c8",
+      "696": "00c8",
+      "697": "8700",
+      "698": "0000",
+      "758": "ff00",
       "978": "ff00",
       "979": "ff00",
       "980": "ff00",
